Add clickable links to images in README.md
I left the original link alone (other than wrapping it in \`) so that it's still easy to select, copy, and paste into the terminal.

I added a hyperlinked :link: emoji so that people can easily "just click it" because some people like to do that.

I also fixed the very strange whitespace. Are you using a monospace font?