create/export: Prefer 'uname -m' to /etc/hostname
In a Fedora VM that I created on my MacBook Air, /etc/hostname is empty,
which means the container's hostname ends with just a period, which
looks odd:

$ cat /etc/hostname

$ ./distrobox enter -n fedora -- cat /etc/hostname
fedora.

Prefer 'uname -n', which shows the host machine's hostname properly.

$ uname -n
fedora

$ ./distrobox enter -n fedora -- uname -n
fedora.fedora

$ ./distrobox enter -n fedora -- cat /etc/hostname
fedora.fedora

uname is described by POSIX and should be available by default on every
modern Linux system, as it is typically a part of the coreutils package.

Link: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/uname.html
Signed-off-by: Nathan Chancellor <nathan@kernel.org>