fix: use container's $PATH when entering (#209)
The PATH fixes in 6b60c73999c5147f732a1f9933d0b52d7d090161 use the
host's $PATH instead of the $PATH defined in the container.  This
change fixes that by inspecting the container and parsing $PATH
out of its environment.  The inspect was already being done so
this shouldn't be a signficiant performance impact.  Future
features may desire further information out of the inspect
invocation and the parsing method can be extended.

The container's $PATH is merged with the host $PATH and the
standard set of binary paths as the existing logic uses.

Co-authored-by: Gerald Britton <gbritton@sekrit.org>