init: fix regression introduced in 3f014de2399885fbfc8b939ab269d78cd5f56c7f
Mounting also /run/systemd/system from host, will trick the systemd
package inside the container to think that the system is booted with
systemd (really just the presence of the path /run/systemd/system will
be sufficient).

Revert the mounting of this path, and document where this is needed.
This also fixes #222