create/init: move some mounts inside init.
Right now the following files are mounted via container manager:
    - /etc/host.conf
    - /etc/hosts
    - /etc/resolv.conf
    - /etc/localtime
Let's do this in the init instead. This ensures that files that are
symlinks, will be refreshed at container restart.

Examples include:
    - changing DNSs
    - changing timezone

This solution will mount the bare minimum to have connectivity, using
the container manager, and then manage all the files from the init.

Signed-off-by: 89luca89 <luca.dimaio1@gmail.com>