init: zypper: enable recommended packages
In openSUSE images, zypper is usually configured to ignore recommended
packages (i.e., weak dependencies). This is one of the reasons why
images are small, and also why only an handful of packages are installed
in the distrobox, during init.

However, this also means that quite a few things (e.g., GUI apps) may
not work out-of-the-box for users, and that more packages would need to
be installed (and it's often not super easy to tell which ones!).

Note also that recommended packages are enabled by default in "regular"
openSUSE/SUSE installs, so having them enabled in the distrobox as well,
is most likely what distrobox users expects.

Let's, therefore, enable them. This makes the initialization of the
dbox (i.e., the first `distrobox enter`) a bit slow, as much more
packages will be installed (and zypper is not really something one can
call fast!), but the improved user experience is worth it.

For the same reason, let's also make sure that we install the docs
files of the RPM packages.

A small side effect is that, on certain image, a package that does
not install properly when inside rootless podman could be pulled in.
Since that is parallel-printer-support, which we'll never going to
need inside a distrobox, we can just lock it out.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>