Add AlmaLinux 9 to tested containers distros
Tested and worked fine. I had to install `dnf-plugins-core` which is
installed out of the box on AlmaLinux 8, but that's not a distrobox
issue.

```shell
distrobox on  main
❯ ./distrobox create -i docker.io/library/almalinux:9 -n alma9 --pre-init-hooks "dnf -y install dnf-plugins-core && dnf config-manager --enable crb && dnf -y install epel-release"
39a00b6fdacdbd22217fba9ff9c08a2468c8a57747f0df6e501ff057231899eb
Distrobox 'alma9' successfully created.
To enter, run:

distrobox-enter alma9

distrobox on  main
❯ ./distrobox enter alma9
Container alma9 is not running.
Starting container alma9
run this command to follow along:

 podman logs -f alma9

 Starting container...                  	[ OK ]
 Executing pre-init hooks...            	[ OK ]
 Installing basic packages...           	[ OK ]
 Setting up read-only mounts...         	[ OK ]
 Setting up read-write mounts...        	[ OK ]
 Setting up host's sockets integration...	[ OK ]
 Integrating host's themes, icons, fonts...	[ OK ]
 Setting up package manager exceptions...	[ OK ]
 Setting up sudo...                     	[ OK ]
 Setting up groups...                   	[ OK ]
 Setting up users...                    	[ OK ]
 Executing init hooks...                	[ OK ]

Container Setup Complete!
Welcome to fish, the friendly interactive shell
Type help for instructions on how to use fish

distrobox on  main
⬢ [almalinux:9] ❯
```

Signed-off-by: Michel Alexandre Salim <michel@michel-slm.name>