Fix gentoo support
this patch includes multiple fixes:
- the local repository no longer gets automatically updated. Other
package managers dont do this either and the output can include
something about an error, that isnt actually a problem, which causes
distrobox-init to fail.
- there are multiple curl packages for gentoo. net-misc/curl is the web
request tool.
- vte-profile doesnt exist for gentoo. The closest package is probably
vte.
- some packages compilation will report errors, which get ignored by the
build tools, so dont matter, but the mention will cause distrobox-init
to abort. To prevent this --quiet-build is used to supress to messages
from compilation