enter: fix the check for container existance
In distrobox enter, we try to check if the container exists. However,
this does not seem to work well, and I always get this, when try to
enter a non-existant container:

    $ distrobox-enter blabla
    Error: error inspecting object: no such container blabla
    Container blabla is not running.
    Starting container blabla
    run this command to follow along:

     podman logs -f blabla

    Error: no container with name or ID "blabla" found: no such container

    An error occurred

Instead of the actual "Cannot find container ..." message.

I'm not sure why the 'eval' logic fails, but let's fix it by just
checking with 'podman ps -a' (or equivalent).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>