enter: offer to automatically create the container
When someone tries to enter a container that does not exist, let's offer
the option of just creating it ourselves. The main (only?) issue is
whether to accept and, if yes, how to handle 'distrobox-create' options
in 'distrobox-enter'.

As a first step, let's ingore the problem entirely and just invoke
'distrobox-create $container_name', in this case.

An improvement could be making 'distrobox-enter' support all
distrobox-create parameter. Of course, that would be "just" the parsing
of them, to stash them somewhere and, if we get to creating the
container, pass them to 'distrobox-create'. That's easy enough to do,
but it then means that every time we add/remove/modify a parameter of
'distrobox-create', we also need to do the same change in
'distrobox-enter' parsing code, which is not really nice...

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>