enter: automatically create containers during 'distrobox-enter' (if they don't exist!) (#302)
* enter: fix the check for container existance

In distrobox enter, we try to check if the container exists. However,
this does not seem to work well, and I always get this, when try to
enter a non-existant container:

    $ distrobox-enter blabla
    Error: error inspecting object: no such container blabla
    Container blabla is not running.
    Starting container blabla
    run this command to follow along:

     podman logs -f blabla

    Error: no container with name or ID "blabla" found: no such container

    An error occurred

Instead of the actual "Cannot find container ..." message.

I'm not sure why the 'eval' logic fails, but let's fix it by just
checking with 'podman ps -a' (or equivalent).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>

* enter: offer to automatically create the container

When someone tries to enter a container that does not exist, let's offer
the option of just creating it ourselves. The main (only?) issue is
whether to accept and, if yes, how to handle 'distrobox-create' options
in 'distrobox-enter'.

As a first step, let's ingore the problem entirely and just invoke
'distrobox-create $container_name', in this case.

An improvement could be making 'distrobox-enter' support all
distrobox-create parameter. Of course, that would be "just" the parsing
of them, to stash them somewhere and, if we get to creating the
container, pass them to 'distrobox-create'. That's easy enough to do,
but it then means that every time we add/remove/modify a parameter of
'distrobox-create', we also need to do the same change in
'distrobox-enter' parsing code, which is not really nice...

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>

* Make interactive/non-interactive mode configurable in config files

Allow users to specify:

non_interactive=true

(or =false, of course), in their config files. This will set the
interactive mode of all the commands, by default, without the need of
passing '-Y'.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>