enter: if flagless container doesn't exist, try it as an image
This is done with a new container variable which can be used as
fallback container_name if it exists otherwise can attempt using it
as image of a new toolbox.

Without a flagless arg, enter tries container_name_default (still my-distrobox)
and then falls back to container_image_default (fedora-toolbox:36).