feat: Replace call to 'rev | cut' with 'grep'
This modifies the parsing of $release_name from $release_ver by
grepping for a semantic version with a '.tar.gz' extension.

The prior method used 'rev | cut | rev' to isolate $release_name
as the last field of $release_ver. The rev tool is not installed
by default on some machines.