Add some checks for fixing issues due to dirs or manpages missing (on openSUSE MicroOS) (#365)
* export: check if the dirs exist before searching them

Dirs in which we look for lanuchers to export may not exist. In this
case, distrobox-export fails like this:

    dario@tumbleweed:~> distrobox-export -a virt-manager
    find: ‘/usr/local/share/applications’: No such file or directory

    An error occurred

Check that they actually exist.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>

* distrobox: handle situations with weird manpages setup

So, assume you're distro strip manpages during packages install. That
might mean the `man` command exists, but there isn't any entry for
anything. In that case, the following happens:

    distrobox --help
    No manual entry for distrobox

    An error occurred

Fix this by calling show_help, if `man` is present but the command
fails.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>