Automatically create docker group matching docker.sock group GID when present.
To ensure optional manual installation of docker in the distrobox can use the docker socket mapped from the host,
it's a necessity that the docker.sock group GID be associated with the container group name 'docker'.
Detect if the socket is present after host socket mapping, and create the 'docker' named group to own it.
Clarify CONTRIBUTING.md for pre-checks