init: Reuse locked flags when mounting read-only
In a user namespace (e.g. when using rootless Podman),
certain mount flags are considered "locked" and cannot be cleared.

The check for these flags is triggered (in this case) when bind-mounting
directories from the host as read-only.

These flags must be explicitly reused when a host directory,
or its nearest mountpoint, uses them.

For example, if /var/lib/flatpak on the host is on a rootfs mounted
using 'nodev' (one of the locked flags), it must be reused.

Signed-off-by: Morgan Kristiansen <morganhjk@gmail.com>