distrobox-export: add login shell option
Add the option (-l or --login) to distrobox export so that the exported
command runs inside a login shell, giving it access to the variables
sourced during login (e.g: /etc/profile).
If sudo is also specified, then care has been taken so that it also runs
in a login shell (via -i).