init: add nvidia integration
This is a preliminary work to implement host's nVidia driver's
integration, in the guest.

The procedure is simple enough, it will try and find all nVidia related
files in the host, and bind mount them in the guest.

Right now this is confirmed working on:

    - Ubuntu 22+
    - Debian 11+
    - Archlinux
    - VoidLinux (glibc)
    - Fedora
    - Redhat/Centos/Family (7+)

This is confirmed NOT working on non-glibc distros like

    - Alpine
    - Void (musl)

This needs more testing and confirmation, I don't have the hardware
sadly.

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>