export: simplify distro_binary script
The current version of the "distrobox_binary" script file creates at
least `1 + 6 x (number of arguments)` extra processes to first quote
and then unquote the arguments, which seems redundant.  If the
arguments contain `'` or `"`, the number of processes would increase
further.  We can simply pass the arguments to the command instead of
quoting the arguments using external commands and then unquote them by
`eval`.

Also, `exec` can be used to run `distribox-enter` in order to reduce
the number of forks.