Add option to disable stats in distrobox-list (saving a lot of time)
This commit adds an option (`--no-stats`) to `distrobox-list`, which,
when set, would not report the cpu and memory usage for the distrobox
containers.

`distrobox-list` spends majority of its time getting the cpu/mem related
stats. On my system (with 6 distrobox containers), the current
`distrbox-list` command takes approximately 4.98 seconds. With this
commit, invoking it as `distrowatch ls --no-stats` only takes 98
miliseonds.