host-exec: Fix CONTAINER_ID: unbound variable error (#888)
* host-exec: Fix CONTAINER_ID: unbound variable error

Currently the script sets the nounset shell option, but it doesn't
define the CONTAINER_ID variable and nor it should (something else does
it). In such cases, we should resort to shell parameter expansion in
order to use a default value (an empty string, in this case) if the
parameter - read CONTAINER_ID variable - is unset or null.

So, we should use "${CONTAINER_ID:-}" instead, to avoid this error.

Fixes #882

Signed-off-by: Luciano Santos <luciano.santos@email.com>

* host-exec: Rename "download" variable

Use "download_command" instead, and initialize it in the top of the
script (download_command=""). This should avoid "unbound variable"
errors in unfortunate cases where curl nor wget are found, and the
variable would remain unset.

Signed-off-by: Luciano Santos <luciano.santos@email.com>

---------

Signed-off-by: Luciano Santos <luciano.santos@email.com>