init/create: setup user password for rootful containers
This commit will create a first `enter` setup for container user, in
case of rootful containers. The default behavious of passwordless
sudo/su is acceptable for rootless containers, where there is no real
privilege escalation, but is a real concern for rootful containers.

This setup will also respect host's sudoers configuration.

Root account will be locked by default in this modality.

During creation there is a really long and difficult flag that will
rever to the old behaviour, **this is highly discouraged!**

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>