debian: pass init if custom image is oldstable now
If you use an old development image that has changed from stable to
oldstable or from oldstable to oldoldstable, plain apt-update will
require you to interactively or explicitly allow it. This patch fixes
both the first run and upgrade scenarios.

Additionally, if you use the vanilla PS1 prompt of a debian install on
the host, debian defaults to prepend ($debian_chroot) to it. Set this to
either CONTAINER_ID or distrobox makes it easier.