export/rm: correctly handle paths containing whitespace and quotes (#1102)
* export: correctly handle paths containing whitespace and quotes

In order to reliably process paths with whitespace or quotes to xargs without using non-POSIX features, each character must be escaped. (Quoting will handle whitespace, but not filenames with quotes in them.)

* rm: correctly handle paths containing whitespace and quotes

In order to reliably process paths with whitespace or quotes to xargs without using non-POSIX features, each character must be escaped. (Quoting will handle whitespace, but not filenames with quotes in them.)