docs: fix hostname matching with --init-hooks
The change #1182 had the intention to evaluate 'uname -n' while
calling 'distrobox-create'. But had the quotation marks wrong.
Double quotation marks must be the outer ones for the evaluation
to occur directly.

When changing /etc/hostname, hostname does not change without a
reboot. To avoid forcing the user to reboot the container after
creating and entering the container, add 'hostname $(uname -n)'
to be called with the --init-hooks flag.

See related issue #1164