init: massively speed up apk deps install
Instead of running "apk info" on each package individually instead just
compare the list of requested packages to an "apk search" of them and
pick the ones that could be found.

On alpine due to the way searching for providers of openssh-client and
exact matching of the found packages (procps is procps-ng) adjust these
both accordingly to keep the end result consistent for alpine and wolfi.

On my end this halves the setup time for the non-toolbox alpine and
wolfi images from ~30 and 40 seconds to ~15 and 20 seconds respectively.