enter: better control over copied host envvars
Currently, all host environment variables are copied into the container,
except for a hardcoded few that are ignored:
https://github.com/89luca89/distrobox/blob/3b9f0e8d3d8bd102e1636a22afffafe00777d30b/distrobox-enter#L434

This is not always the desired behavior, particularly on host distros
like NixOS or Guix that break from the FHS and need to set a lot of
environment variables to work around the resulting issues:
https://github.com/89luca89/distrobox/issues/1516

Therefore, provide a --no-envvars option to disable the default copying
of environment variables, and an --additional-envvars option to copy
specific ones.

It was suggested to use `env -u distrobox` in order to unset specific
environment variables:
https://github.com/89luca89/distrobox/issues/656
However, there are usually too many copied environment variables to
unset them all individually.
`env -i distrobox` will unset environment variables, but that includes
PATH and other things needed for distrobox to execute correctly on the
host.

Other instances of this feature being requested:
https://github.com/89luca89/distrobox/issues/508
https://github.com/89luca89/distrobox/issues/743
https://github.com/89luca89/distrobox/discussions/1173