enter: don't filter environment variables with special characters
Previously when passing environment variables from the host to the
container, the script filtered out any variables that contained special
characters. This could lead to important environment variables being
omitted when entering the container.
In this commit, I modified it to only pass the names of the environment
variables to the --env flag, rather than their full key-value pairs, and
let container runtime read the values.
I also changed printenv to separate lines with null characters to avoid
issues with newlines in variable values.