feat(ephemeral): implement ephemeral command (#34)
* refactor(enter): define command options

Refactor the `enter` command package so that adheres to the convention
to define its own arguments as options and read them in the Execute
method.

* refactor(containermanager): add InspectContainer to the public interface

* refactor(rm): move preflight checks in the command layer

That simplifies the containermanager.Remove interface and improve the
separation between the container manager and the presentation logic as
prompting it's not handled at containermanager level anymore.
It also improves efficiency as the userenv is fetch only once.

* refactor(ui): pass the prompter by reference

Use pointer to optimize memory footprint. It also allows nullable
references, hence a validation is required for the `rm` command when
executing on interactive mode.

* feat(ephemeral): add ephemeral command

The command is a composition of create, enter and rm commands.
It accepts the same arguments as the create command.